// src/groups/item/sites/item/operations/item/richLongRunningOperationItemRequestBuilderDeleteRequestConfiguration.ts
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface RichLongRunningOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}